+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 - Added miliseconds (optional) to logs
 - Import_webFile (download a file to a local folder and make it available to the 'new' function)
 - Bug fixes
+- Auto name for objects. You can create object without specify its name (a random name will be generated and returned in the _r variable)
+- Helper (in utils.sh) to work with 'named arguments' in functions (allow calls like 'my_function "arg=value" "arg2=value2"')
+- Object finalization/free function in the 'new.sh'